Dedicated /analysis page per analysis
Each stored analysis now has its own URL at /analysis?id=<uuid>
instead of expanding inline inside the past-analyses list on
/analyze. Cleaner focus (one analysis per page), better for
sharing internal links, mobile-friendlier.

Page structure
- New src/app/(app)/analysis/page.tsx reads ?id= and fetches from
  the user's history hook (RLS already scopes to own rows). Shows:
  - Breadcrumb back to /analyze
  - Filename + full metadata header (role / level / event / stage
    / tags / dancer-focus / created-at)
  - Action bar: Watch (auto-loads if video stored), Re-analyze,
    Delete video, Share / Copy / Stop sharing, Delete analysis.
    View count badge with last-viewed tooltip.
  - Timeline + video (auto-loaded when a stored clip exists —
    dedicated page has room for it without a click)
  - Full ScoreResultCard (score hero, categories, partner cards,
    pattern summary, strengths, improvements)
- Wrapped in Suspense because useSearchParams requires it.
- Graceful error states: no id, analysis not found.

HistoryRow simplified
- Was ~470 lines with inline expand/collapse + all action logic.
- Now ~100 lines — just a Link-wrapped compact summary card.
  Click navigates to /analysis?id=<uuid>. All action UI moved to
  the dedicated page so there's one canonical place to interact
  with an analysis.
- Kept autoExpand prop API for now (ring-highlight on recent
  uploads still works, though the row itself no longer expands).
- Surface "stored" and "shared" indicators as small icons in the
  metadata row so users can tell at a glance which clips are
  replayable and which links are live.

Shared components
- Extracted ScoreResultCard + ScoreBar + TechniqueBreakdown +
  PartnerCards + PartnerPanel + related helpers from analyze/
  page.tsx into src/components/analyze/score-result-card.tsx.
  Both /analyze (fresh-run results) and /analysis (stored
  results) import from there.

Deep-links updated
- Dashboard recent-analyses links → /analysis?id=<id>
- Chart onOpen → /analysis?id=<id>
- Any /analyze?id=<id> URLs will still work but just open /analyze
  with an auto-expand target; users arriving from old links land
  on the list, which is acceptable.
